# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Training Programme Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface WeekProgression {
  week: number;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: string | null;
  /** @nullable */
  rpe?: string | null;
  /** @nullable */
  weight?: string | null;
}

export interface Exercise {
  id: string;
  name: string;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: string | null;
  /** @nullable */
  rpe?: string | null;
  /** @nullable */
  rest?: string | null;
  /** @nullable */
  tempo?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  rawText?: string | null;
  weekProgression?: WeekProgression[];
}

export interface Programme {
  id: number;
  title: string;
  exercises: Exercise[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateProgrammeBody {
  title: string;
  exercises: Exercise[];
}

export interface UpdateProgrammeBody {
  title?: string;
  exercises?: Exercise[];
}

export interface ParseTranscriptBody {
  transcript: string;
  existingExercises?: Exercise[];
}

export interface ParseTranscriptResponse {
  exercises: Exercise[];
  rawTranscript: string;
}

export interface TranscribeResponse {
  transcript: string;
}

export type TranscribeAudioBody = {
  audio: Blob;
};