# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Training Programme Builder API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all programmes
 */
export const ListProgrammesResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  exercises: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      sets: zod.number().nullish(),
      reps: zod.string().nullish(),
      rpe: zod.string().nullish(),
      rest: zod.string().nullish(),
      tempo: zod.string().nullish(),
      notes: zod.string().nullish(),
      rawText: zod.string().nullish(),
      weekProgression: zod
        .array(
          zod.object({
            week: zod.number(),
            sets: zod.number().nullish(),
            reps: zod.string().nullish(),
            rpe: zod.string().nullish(),
            weight: zod.string().nullish(),
          }),
        )
        .optional(),
    }),
  ),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListProgrammesResponse = zod.array(ListProgrammesResponseItem);

/**
 * @summary Create a new programme
 */
export const CreateProgrammeBody = zod.object({
  title: zod.string(),
  exercises: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      sets: zod.number().nullish(),
      reps: zod.string().nullish(),
      rpe: zod.string().nullish(),
      rest: zod.string().nullish(),
      tempo: zod.string().nullish(),
      notes: zod.string().nullish(),
      rawText: zod.string().nullish(),
      weekProgression: zod
        .array(
          zod.object({
            week: zod.number(),
            sets: zod.number().nullish(),
            reps: zod.string().nullish(),
            rpe: zod.string().nullish(),
            weight: zod.string().nullish(),
          }),
        )
        .optional(),
    }),
  ),
});

/**
 * @summary Get a programme by ID
 */
export const GetProgrammeParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProgrammeResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  exercises: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      sets: zod.number().nullish(),
      reps: zod.string().nullish(),
      rpe: zod.string().nullish(),
      rest: zod.string().nullish(),
      tempo: zod.string().nullish(),
      notes: zod.string().nullish(),
      rawText: zod.string().nullish(),
      weekProgression: zod
        .array(
          zod.object({
            week: zod.number(),
            sets: zod.number().nullish(),
            reps: zod.string().nullish(),
            rpe: zod.string().nullish(),
            weight: zod.string().nullish(),
          }),
        )
        .optional(),
    }),
  ),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update a programme
 */
export const UpdateProgrammeParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProgrammeBody = zod.object({
  title: zod.string().optional(),
  exercises: zod
    .array(
      zod.object({
        id: zod.string(),
        name: zod.string(),
        sets: zod.number().nullish(),
        reps: zod.string().nullish(),
        rpe: zod.string().nullish(),
        rest: zod.string().nullish(),
        tempo: zod.string().nullish(),
        notes: zod.string().nullish(),
        rawText: zod.string().nullish(),
        weekProgression: zod
          .array(
            zod.object({
              week: zod.number(),
              sets: zod.number().nullish(),
              reps: zod.string().nullish(),
              rpe: zod.string().nullish(),
              weight: zod.string().nullish(),
            }),
          )
          .optional(),
      }),
    )
    .optional(),
});

export const UpdateProgrammeResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  exercises: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      sets: zod.number().nullish(),
      reps: zod.string().nullish(),
      rpe: zod.string().nullish(),
      rest: zod.string().nullish(),
      tempo: zod.string().nullish(),
      notes: zod.string().nullish(),
      rawText: zod.string().nullish(),
      weekProgression: zod
        .array(
          zod.object({
            week: zod.number(),
            sets: zod.number().nullish(),
            reps: zod.string().nullish(),
            rpe: zod.string().nullish(),
            weight: zod.string().nullish(),
          }),
        )
        .optional(),
    }),
  ),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a programme
 */
export const DeleteProgrammeParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Duplicate a programme
 */
export const DuplicateProgrammeParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Parse a voice transcript into structured exercises
 */
export const ParseTranscriptBody = zod.object({
  transcript: zod.string(),
  existingExercises: zod
    .array(
      zod.object({
        id: zod.string(),
        name: zod.string(),
        sets: zod.number().nullish(),
        reps: zod.string().nullish(),
        rpe: zod.string().nullish(),
        rest: zod.string().nullish(),
        tempo: zod.string().nullish(),
        notes: zod.string().nullish(),
        rawText: zod.string().nullish(),
        weekProgression: zod
          .array(
            zod.object({
              week: zod.number(),
              sets: zod.number().nullish(),
              reps: zod.string().nullish(),
              rpe: zod.string().nullish(),
              weight: zod.string().nullish(),
            }),
          )
          .optional(),
      }),
    )
    .optional(),
});

export const ParseTranscriptResponse = zod.object({
  exercises: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      sets: zod.number().nullish(),
      reps: zod.string().nullish(),
      rpe: zod.string().nullish(),
      rest: zod.string().nullish(),
      tempo: zod.string().nullish(),
      notes: zod.string().nullish(),
      rawText: zod.string().nullish(),
      weekProgression: zod
        .array(
          zod.object({
            week: zod.number(),
            sets: zod.number().nullish(),
            reps: zod.string().nullish(),
            rpe: zod.string().nullish(),
            weight: zod.string().nullish(),
          }),
        )
        .optional(),
    }),
  ),
  rawTranscript: zod.string(),
});

/**
 * @summary Transcribe audio to text
 */
export const TranscribeAudioBody = zod.object({
  audio: zod.instanceof(File),
});

export const TranscribeAudioResponse = zod.object({
  transcript: zod.string(),
});