# lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Training Programme Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { WeekProgression } from "./weekProgression";

export interface Exercise {
  id: string;
  name: string;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: string | null;
  /** @nullable */
  rpe?: string | null;
  /** @nullable */
  rest?: string | null;
  /** @nullable */
  tempo?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  rawText?: string | null;
  weekProgression?: WeekProgression[];
}