# lib/api-zod/src/generated/types/parseTranscriptBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Training Programme Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from "./exercise";

export interface ParseTranscriptBody {
  transcript: string;
  existingExercises?: Exercise[];
}