# lib/api-zod/src/generated/types/programme.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Training Programme Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from "./exercise";

export interface Programme {
  id: number;
  title: string;
  exercises: Exercise[];
  createdAt: Date;
  updatedAt: Date;
}