# lib/api-zod/src/generated/types/weekProgression.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Training Programme Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface WeekProgression {
  week: number;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: string | null;
  /** @nullable */
  rpe?: string | null;
  /** @nullable */
  weight?: string | null;
}